pyproject.toml
[tool.poetry]
name = "webexbot-langgraph-ai-assistant"
version = "0.1.0"
description = ""
authors = ["evfedoto"]
readme = "README.md"
packages = [{ include = "src" }, { include = "src/**/*.py" }]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.35"
langchain = "^0.3.3"
langchain-openai = "^0.2.2"
python-dotenv = "^1.0.1"
langgraph-checkpoint-sqlite = "^2.0.0"
webex-bot = "^0.5.2"
langchain-community = "^0.3.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.8"
mypy = "^1.11.2"
poethepoet = "^0.29.0"
coverage = "^7.6.3"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"

[tool.poe.tasks]
_rf = "poetry run ruff format ."
_rc = "poetry run ruff check . --fix"
_mypy = "poetry run mypy ."
format = ["_rf", "_rc", "_mypy"]
update = "poetry update"
start = "poetry run python -m webexbot.webexbot"
test = "poetry run pytest --cov --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
